ajout prix !!! enfin

+++ front/js/cart.js
@@ -1,23 +1,39 @@
 let productLocalStorage=JSON.parse(localStorage.getItem("product"));//Json vers javascript
 let cartItem= document.getElementById("cart__items")// récupère la div cart__items
-console.log(productLocalStorage)
-for (i=0;i < productLocalStorage.length; i++){//parcour le tableau product dans le local storage
+let kanapInfos = JSON.parse(localStorage.getItem("product"));
+
+if (kanapInfos) {
+    for (let product of kanapInfos) {
+        let item = {
+            Id: product.productId,
+            color: product.colorValue,
+            quantity: product.quantityProduct,
+        }
+         
+        fetch("http://localhost:3000/api/products/" + item.Id)
+            .then(function(res) {
+                if (res.ok) {
+                    return res.json();
+                }
+            })
+        .then(function(product) {
+//parcour le tableau product dans le local storage
 //constante pour l'affichage du panier via les éléments du local storage   
     const PRODUCTCART=`
-    <article class='cart__item' data-id='${productLocalStorage[i].productId}' data-color='${productLocalStorage[i].colorValue}'>
+    <article class='cart__item' data-id='${item.Id}' data-color='${item.color}'>
     <div class='cart__item__img'>
-      <img src='${productLocalStorage[i].imgProduct}' alt='${productLocalStorage[i].altProduct}'>
+      <img src='${product.imageUrl}' alt='${product.altTxt}'>
     </div>
     <div class='cart__item__content'>
       <div class='cart__item__content__description'>
-        <h2>${productLocalStorage[i].productName}</h2>
-        <p>Couleur ${productLocalStorage[i].colorValue}</p>
-        <p>${productLocalStorage[i].price}€</p>
+        <h2>${product.name}</h2>
+        <p>Couleur ${item.color} </p>
+        <p>${product.price}€</p>
       </div>
       <div class='cart__item__content__settings'>
         <div class='cart__item__content__settings__quantity'>
           <p>Qté :</p>
-          <input type='number' class='itemQuantity' name='itemQuantity' min='1' max='100' value=${productLocalStorage[i].quantityProduct}>
+          <input type='number' class='itemQuantity' name='itemQuantity' min='1' max='100' value=${item.quantity}>
         </div>
         <div class='cart__item__content__settings__delete'>
           <p class='deleteItem'>Supprimer</p>
@@ -26,7 +42,7 @@ for (i=0;i < productLocalStorage.length; i++){//parcour le tableau product dans 
     </div>
   </article>`
 cartItem.innerHTML +=PRODUCTCART;//Ajout des constant HTML
-}
+})}}
 /*Changement de quantité*/
 let inputQuantity=document.querySelectorAll(".itemQuantity")// récupère la class dans le dom
 for (let i =0; i<inputQuantity.length; i++){//on fait une boucle de tous les itemQuantité
